+++ gemma-core/src/main/java/ubic/gemma/model/analysis/expression/diff/DifferentialExpressionAnalysisValueObject.java
@@ -16,6 +16,7 @@ package ubic.gemma.model.analysis.expression.diff;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.Hibernate;
@@ -50,11 +51,18 @@ public class DifferentialExpressionAnalysisValueObject extends AnalysisValueObje
     private Collection<DiffExResultSetSummaryValueObject> resultSets = new HashSet<>();
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Collection<Long> arrayDesignsUsed;
-    private Long bioAssaySetId;
-    private Long sourceExperiment;
+    private Long experimentAnalyzedId;
+    private Long sourceExperimentId;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
     private ExperimentalFactorValueObject subsetFactor;
+    @JsonInclude(JsonInclude.Include.NON_NULL)
     private Long subsetFactorId;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
     private FactorValueValueObject subsetFactorValue;
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private Long subsetFactorValueId;
 
     public DifferentialExpressionAnalysisValueObject() {
         super();
@@ -67,21 +75,44 @@ public class DifferentialExpressionAnalysisValueObject extends AnalysisValueObje
      */
     public DifferentialExpressionAnalysisValueObject( DifferentialExpressionAnalysis analysis ) {
         super( analysis );
-        this.bioAssaySetId = analysis.getExperimentAnalyzed().getId();
+        this.experimentAnalyzedId = analysis.getExperimentAnalyzed().getId();
         // experimentAnalyzed is eagerly fetched
         if ( analysis.getExperimentAnalyzed() instanceof ExpressionExperimentSubSet ) {
             // sourceExperiment is eagerly fetched too
-            this.sourceExperiment = ( ( ExpressionExperimentSubSet ) analysis.getExperimentAnalyzed() ).getSourceExperiment().getId();
+            this.sourceExperimentId = ( ( ExpressionExperimentSubSet ) analysis.getExperimentAnalyzed() ).getSourceExperiment().getId();
         }
-        if ( analysis.getSubsetFactorValue() != null && Hibernate.isInitialized( ( analysis.getSubsetFactorValue() ) ) ) {
-            this.subsetFactorValue = new FactorValueValueObject( analysis.getSubsetFactorValue() );
-            this.subsetFactorId = analysis.getSubsetFactorValue().getExperimentalFactor().getId();
-            if ( Hibernate.isInitialized( analysis.getSubsetFactorValue().getExperimentalFactor() ) ) {
-                this.subsetFactor = new ExperimentalFactorValueObject(
-                        analysis.getSubsetFactorValue().getExperimentalFactor() );
+        if ( analysis.getSubsetFactorValue() != null ) {
+            if ( Hibernate.isInitialized( analysis.getSubsetFactorValue() ) ) {
+                this.subsetFactorValue = new FactorValueValueObject( analysis.getSubsetFactorValue(), false );
+                if ( Hibernate.isInitialized( analysis.getSubsetFactorValue().getExperimentalFactor() ) ) {
+                    this.subsetFactor = new ExperimentalFactorValueObject(
+                            analysis.getSubsetFactorValue().getExperimentalFactor() );
+                } else {
+                    this.subsetFactorId = analysis.getSubsetFactorValue().getExperimentalFactor().getId();
+                }
+            } else {
+                this.subsetFactorValueId = analysis.getSubsetFactorValue().getId();
             }
             // fill in the factorValuesUsed separately, needs access to details of the subset.
         }
+    }
+
+    /**
+     * @deprecated use {@link #getExperimentAnalyzedId()} instead
+     */
+    @Deprecated
+    @Schema(description = "This is deprecated, use experimentAnalyzedId instead.", deprecated = true)
+    public Long getBioAssaySetId() {
+        return experimentAnalyzedId;
+    }
+
+    /**
+     * @deprecated use {@link #getSourceExperimentId()} instead
+     */
+    @Deprecated
+    @Schema(description = "This is deprecated, use sourceExperimentId instead.", deprecated = true)
+    public Long getSourceExperiment() {
+        return sourceExperimentId;
     }
 
     /**
@@ -103,17 +134,18 @@ public class DifferentialExpressionAnalysisValueObject extends AnalysisValueObje
      * This can be null in certain cases if set to NULL via {@link #setFactorValuesUsed(Map)} so that it does not appear
      * in the JSON serialization, but you can assume it is non-null.
      */
+    @JsonInclude(JsonInclude.Include.NON_NULL)
     public Map<Long, Collection<FactorValueValueObject>> getFactorValuesUsedByExperimentalFactorId() {
         return factorValuesUsed;
     }
 
     @JsonProperty("isSubset")
     public boolean isSubset() {
-        return this.subsetFactor != null;
+        return this.sourceExperimentId != null || this.subsetFactorValue != null || this.subsetFactorValueId != null;
     }
 
     @Override
     public String toString() {
-        return "DiffExAnalysisVO [id=" + id + ", bioAssaySetId=" + bioAssaySetId + "]";
+        return "DiffExAnalysisVO [id=" + id + ", experimentAnalyzedId=" + experimentAnalyzedId + "]";
     }
 }